requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1
langchain==0.1.16
langchain-google-genai==0.0.11
google-generativeai==0.4.1
